Convert use-focus-on-blur to TS
